refactor: main window vue component extract
