+++ src/pages/main/App.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 // This starter template is using Vue 3 <script setup> SFCs
 // Check out https://vuejs.org/api/sfc-script-setup.html#script-setup
-import SnapVault from '~/components/SnapVault.vue'
+import SnapVault from '~/pages/main/components/SnapVault.vue'
 </script>
 
 <template>
