+++ src/pages/main/components/FileInfo.vue
