+++ src/pages/main/components/SnapVault.vue
