+++ src/pages/main/components/SnapVaultItem.vue
